requirements.txt
streamlit==1.39.0
pandas==2.2.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
faker==25.8.0
plotly==5.24.1
